Create Utils class and move several methods not directly related to WFC logic
